retoques de funciones y relaciones

+++ src/main/java/com/TFG/TFG/Respository/CategoryRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CategoryRepository extends JpaRepository<Category,Long> {
-    Category findById(long id);
+public interface CategoryRepository extends JpaRepository<Category,Integer> {
+    Category findById(int id);
 }
