+++ src/main/java/com/TFG/TFG/Respository/ReviewRepository.java
@@ -5,9 +5,9 @@ import com.TFG.TFG.Model.Review;
 import com.TFG.TFG.Model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ReviewRepository extends JpaRepository<Review,Long> {
+public interface ReviewRepository extends JpaRepository<Review,Integer> {
 
-    Review findById(long id);
+    Review findById(int id);
     Review findByProduct(Producto producto);
     Review findByUser(User user);
 }
